Rename the benchmarks and the tests for the multi producer multi consumer hashtable in test/bench=hashtable-mpmc-*
