+++ benches/bench-hashtable-mpmc-op-get.cpp
@@ -18,7 +18,7 @@
 #include "data_structures/hashtable/mcmp/hashtable_op_get.h"
 
 #include "../tests/support.h"
-#include "../tests/hashtable/fixtures-hashtable.h"
+#include "hashtable/mpmc/fixtures-hashtable-mpmc.h"
 
 #include "benchmark-program.hpp"
 #include "benchmark-support.hpp"
@@ -120,12 +120,14 @@ static void hashtable_op_get_single_key_external(benchmark::State& state) {
         bucket_index = test_key_1_hash % hashtable->ht_current->buckets_count;
         chunk_index = HASHTABLE_TO_CHUNK_INDEX(bucket_index);
         chunk_slot_index = 0;
+        char *test_key_1_clone = (char*)slab_allocator_mem_alloc_zero(test_key_1_len + 1);
+        strncpy(test_key_1_clone, test_key_1, test_key_1_len);
 
         HASHTABLE_SET_KEY_EXTERNAL_BY_INDEX(
                 chunk_index,
                 chunk_slot_index,
                 test_key_1_hash,
-                test_key_1,
+                test_key_1_clone,
                 test_key_1_len,
                 test_value_1);
     }
@@ -160,11 +162,7 @@ static void hashtable_op_get_single_key_external(benchmark::State& state) {
 
 static void BenchArguments(benchmark::internal::Benchmark* b) {
     b
-            ->ArgsProduct({
-                                  { 0x0000FFFFu, 0x000FFFFFu, 0x001FFFFFu, 0x007FFFFFu, 0x00FFFFFFu, 0x01FFFFFFu, 0x07FFFFFFu,
-                                          0x0FFFFFFFu, 0x1FFFFFFFu, 0x3FFFFFFFu, 0x7FFFFFFFu },
-                                  { 50, 75 },
-                          })
+            ->Arg(256)
             ->ThreadRange(TEST_THREADS_RANGE_BEGIN, TEST_THREADS_RANGE_END)
             ->Iterations(10000000)
             ->DisplayAggregatesOnly(false);
