+++ tests/hashtable/mpmc/fixtures-hashtable-mpmc.h
@@ -1,5 +1,5 @@
-#ifndef CACHEGRAND_FIXTURES_HASHTABLE_H
-#define CACHEGRAND_FIXTURES_HASHTABLE_H
+#ifndef CACHEGRAND_FIXTURES_HASHTABLE_MPMC_H
+#define CACHEGRAND_FIXTURES_HASHTABLE_MPMC_H
 
 #define HASHTABLE_SUPPORT_HASH_SEED    42U
 
@@ -158,4 +158,4 @@ hashtable_hash_quarter_t test_key_long_1_hash_quarter = test_key_long_1_hash_hal
 }
 #endif
 
-#endif //CACHEGRAND_FIXTURES_HASHTABLE_H
+#endif //CACHEGRAND_FIXTURES_HASHTABLE_MPMC_H
