+++ tests/hashtable/mpmc/test-hashtable-mpmc-op-delete.cpp
@@ -21,8 +21,8 @@
 #include "data_structures/hashtable/mcmp/hashtable_op_set.h"
 #include "data_structures/hashtable/mcmp/hashtable_op_delete.h"
 
-#include "../support.h"
-#include "fixtures-hashtable.h"
+#include "../../support.h"
+#include "fixtures-hashtable-mpmc.h"
 
 TEST_CASE("hashtable/hashtable_mcmp_op_delete.c", "[hashtable][hashtable_op][hashtable_mcmp_op_delete]") {
     SECTION("hashtable_mcmp_op_delete") {
