+++ tests/hashtable/mpmc/test-hashtable-mpmc-op-get.cpp
@@ -24,8 +24,8 @@
 #include "data_structures/hashtable/mcmp/hashtable_support_index.h"
 #include "data_structures/hashtable/mcmp/hashtable_op_get.h"
 
-#include "../support.h"
-#include "fixtures-hashtable.h"
+#include "../../support.h"
+#include "fixtures-hashtable-mpmc.h"
 
 TEST_CASE("hashtable/hashtable_mcmp_op_get.c", "[hashtable][hashtable_op][hashtable_mcmp_op_get]") {
     SECTION("hashtable_mcmp_op_get") {
