+++ tests/hashtable/mpmc/test-hashtable-mpmc-op-iter.cpp
@@ -24,8 +24,8 @@
 #include "data_structures/hashtable/mcmp/hashtable_support_index.h"
 #include "data_structures/hashtable/mcmp/hashtable_op_iter.h"
 
-#include "../support.h"
-#include "fixtures-hashtable.h"
+#include "../../support.h"
+#include "fixtures-hashtable-mpmc.h"
 
 TEST_CASE("hashtable/hashtable_mcmp_op_iter.c", "[hashtable][hashtable_op][hashtable_mcmp_op_iter]") {
     SECTION("hashtable_mcmp_op_iter") {
