+++ tests/hashtable/mpmc/test-hashtable-mpmc-support-index.cpp
@@ -15,7 +15,7 @@
 #include "data_structures/hashtable/mcmp/hashtable.h"
 #include "data_structures/hashtable/mcmp/hashtable_support_index.h"
 
-#include "fixtures-hashtable.h"
+#include "fixtures-hashtable-mpmc.h"
 
 #if CACHEGRAND_CMAKE_CONFIG_USE_HASH_ALGORITHM_T1HA2 == 1
 hashtable_bucket_index_t test_key_1_hash_buckets_0x80 = 76;
